API300 NetworkSet Fixed attributes conflict
