+++ libraries/resources/api300/c7000/network_set_provider.rb
@@ -1,0 +1,10 @@
+module OneviewCookbook
+  module API300
+    module C7000
+      # NetworkSet Provider resource methods
+      module NetworkSetProvider
+        include OneviewCookbook::API200::NetworkSetProvider
+      end
+    end
+  end
+end
