+++ libraries/resources/api300/synergy/network_set_provider.rb
@@ -1,0 +1,10 @@
+module OneviewCookbook
+  module API300
+    module Synergy
+      # NetworkSet Provider resource methods
+      module NetworkSetProvider
+        include OneviewCookbook::API300::C7000::NetworkSetProvider
+      end
+    end
+  end
+end
